+++ java/ab/j3d/model/Node3D.java
@@ -238,7 +238,7 @@ public class Node3D
 	 * @param   fillColor       Color to use for filling faces (<code>null</code> to disable drawing).
 	 * @param   shadeFactor     Amount of shading that may be applied (0=none, 1=extreme).
 	 */
-	public void paint( final Graphics2D g , final Matrix3D gTransform , final Matrix3D viewTransform , final Color outlineColor , final Color fillColor , final float shadeFactor )
+	public void paint( final Graphics2D g , final Matrix3D gTransform , final Matrix3D viewTransform , final Color outlineColor , final Color fillColor , final double shadeFactor )
 	{
 		final Node3D[] children = getChildren();
 		for ( int i = 0 ; i < children.length ; i++ )
